% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance.R
\name{anova.default}
\alias{anova.default}
\title{Default method for computing analysis of variance/deviance.}
\usage{
\method{anova}{default}(object, ...)
}
\arguments{
\item{object}{a model object.}

\item{...}{additional arguments passed to specific methods.}
}
\value{
a data frame with the analysis results.
}
\description{
Performs an analysis of variance/deviance for model objects as described
for the \code{\link[stats]{anova}} generic. In addition, percent of explained variance/deviance is returned.
}
