% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance.R
\name{anova.lm_analysis}
\alias{anova.lm_analysis}
\title{Analysis of variance or deviance table for lm_analysis objects.}
\usage{
\method{anova}{lm_analysis}(object, ...)
}
\arguments{
\item{object}{an object of class 'lm_analysis' created e.g. by \code{\link{make_lm}}.}

\item{...}{extra arguments passed to model-specific methods like \code{\link[stats]{anova.lm}}
or \code{\link[stats]{anova.glm}}.}
}
\value{
a data frame with ANOVA or deviance analysis table, fraction of explained variance/deviance
and, optionally, test results.
}
\description{
Performs analysis of variance (lm) or deviance (glm) for lm_analysis objects.
}
\details{
The terms are added sequentially, see: \code{\link[stats]{anova.lm}}, \code{\link[stats]{anova.glm}} or
\code{\link[mgcv]{anova.gam}} for details and additional arguments.
The fraction of explained variance is based on sum of squares for lm class objects or on deviance for glm models.
This parameter is currently absent from the output for GAM models. Polr models return NULL and a warning
}
