% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_expected_}
\alias{calc_expected_}
\title{Calculate expected normal values for the given variable.}
\usage{
calc_expected_(data, observed)
}
\arguments{
\item{data}{data frame.}

\item{observed}{name of the variable of interest}
}
\value{
A data frame with the extra variable .expect.norm with the expected normal distribution values
}
\description{
Calculate expected normal values for the given variable.
}
\details{
credits to: https://stackoverflow.com/questions/43217104/coloring-points-in-a-geom-qq-plot
}
