% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_funs.R
\name{count_model}
\alias{count_model}
\title{Extract variable n numbers.}
\usage{
count_model(linear_model)
}
\arguments{
\item{linear_model}{lm, glm, polr or gam class model or an object of the lm_analysis class.}
}
\value{
a data frame with counts
}
\description{
extracts the numbers of complete cases for numeric features and level counts for categorical variables.
}
