% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_funs.R
\name{distribution}
\alias{distribution}
\title{Distribution of the model response}
\usage{
distribution(
  linear_model,
  facet_formula = NULL,
  cust_theme = ggplot2::theme_classic(),
  ...
)
}
\arguments{
\item{linear_model}{lm, glm, polr or lm_analysis class model.}

\item{facet_formula}{formula for faceting the histogram plots. If NULL, no faceting is done.}

\item{cust_theme}{customized plot theme provided by the user.}

\item{...}{additional arguments passed to \code{\link[ggplot2]{geom_histogram}}.}
}
\value{
a ggplot histogram.
}
\description{
Plots a histogram of the model response or a series of histograms if the faceting formula argument
is provided by the user, see: \code{\link[ggplot2]{facet_grid}} for details.
}
