% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop.R
\name{extract.lm_analysis}
\alias{extract.lm_analysis}
\title{Extract features of lm_analysis objects}
\usage{
\method{extract}{lm_analysis}(
  lm_analysis_object,
  what = c("data", "formula", "inference", "fit", "n", "n_levels"),
  ...
)
}
\arguments{
\item{lm_analysis_object}{an object of class 'lm_analysis' created e.g. by \code{\link{make_lm}}.}

\item{what}{name of the requested feature.}

\item{...}{extra parameters passed to inference (\code{\link{get_estimates}}), fit (\code{\link{get_stats}})
and prediction functions (\code{\link{predict.lm_analysis}})}
}
\value{
the requested feature.
}
\description{
a handy extractor function enabling access to the model frame, formulas and more.
}
\details{
'data' returns the model frame, 'formula' returns the formula, 'inference' returns
the inference summary, 'fit' the summary of fit stats, 'n' the number of complete cases, 'n_levels' the number
of complete observations within the variable levels
}
