% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop.R
\name{extract}
\alias{extract}
\title{Extract features of an object}
\usage{
extract(x, ...)
}
\value{
The requested feature.
}
\description{
S3 generic functions used to extract object-specific properties or other features
}
