% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop.R
\name{formula.lm_analysis}
\alias{formula.lm_analysis}
\title{Formula of a lm_analysis object}
\usage{
\method{formula}{lm_analysis}(x, ...)
}
\arguments{
\item{lm_analysis_object}{an object of class 'lm_analysis' created e.g. by \code{\link{make_lm}}.}
}
\value{
the formula used for modeling.
}
\description{
The model formula of a lm_analysis object as specified by \code{\link[stats]{formula}}.
}
