% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_funs.R
\name{get_qc_plots}
\alias{get_qc_plots}
\title{Diagnostic plots of model residuals}
\usage{
get_qc_plots(linear_model, cust_theme = ggplot2::theme_classic(), ...)
}
\arguments{
\item{linear_model}{lm, glm, polr or lm_analysis class model.}

\item{cust_theme}{customized plot theme provided by the user.}

\item{...}{extra arguments passed to \code{\link[broom]{augment}} and \code{\link{get_qc_tbl}}.}
}
\value{
returns a series of ggplot objects with the diagnostic residuals plots.
}
\description{
Plots a series of diagnostic plots of model residuals with potential outliers indicated.
}
