% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_funs.R
\name{homogeneity}
\alias{homogeneity}
\title{Variance homogeneity of the model residuals}
\usage{
homogeneity(linear_model, ...)
}
\arguments{
\item{linear_model}{lm, glm, polr or lm_analysis class model.}

\item{...}{extra arguments passed to \code{\link[broom]{augment}} and \code{\link{get_qc_tbl}}.}
}
\value{
a tibble with the test results: test statistic, degrees of freedom and p value.
}
\description{
Tests homogeneity of the model residuals by Levene test (\code{\link[DescTools]{LeveneTest}}).
}
