% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_constructors.R
\name{is_lm_analysis}
\alias{is_lm_analysis}
\title{Check lm_analysis class.}
\usage{
is_lm_analysis(x)
}
\arguments{
\item{x}{an object.}
}
\value{
a logical value.
}
\description{
Checks if the object is an instance of the lm_analysis class.
}
