% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_funs.R
\name{linearity}
\alias{linearity}
\title{Graphical inspection of the linearity assumption}
\usage{
linearity(linear_model, cust_theme = ggplot2::theme_classic())
}
\arguments{
\item{linear_model}{lm, glm, polr or lm_analysis class model.}

\item{cust_theme}{customized plot theme provided by the user.}
}
\value{
a list of ggplot objects with point plots and fitted LOESS trends.
}
\description{
Plots the relationship of the numeric model response and the explanatory factors.
}
