% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_constructors.R
\name{make_lm}
\alias{make_lm}
\title{Construct a lm or glm analysis object.}
\usage{
make_lm(
  data,
  response = NULL,
  indep_variable = NULL,
  formula = NULL,
  confounder = NULL,
  weight_variable = NULL,
  mod_fun = glm,
  family = "binomial",
  error_resistant = TRUE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{data}{data frame.}

\item{response}{name of the outcome variable, ignored if 'formula' provided.}

\item{indep_variable}{name of the explanatory variable or a vector with explanatory variable names,
ignored if 'formula' provided.}

\item{confounder}{name of a confounding variable, treated as such by the downstream methods.}

\item{weight_variable}{name of a model weight variable.}

\item{mod_fun}{modeling function. Currently supported are lm, glm, MASS::polr and mgcv::gam.}

\item{family}{modeling function family}

\item{error_resistant}{logical, if TRUE, NULL is returned without an error but a warning is raised}

\item{verbose}{logical, should the}

\item{...}{extra arguments passed to the modeling function}
}
\value{
An object of class 'lm_analysis'
}
\description{
Builds a linear or generalized linear model, optionally in an error-resistant mode.
}
