% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop.R
\name{model.frame.lm_analysis}
\alias{model.frame.lm_analysis}
\title{Model frame of a lm_analysis object}
\usage{
\method{model.frame}{lm_analysis}(lm_analysis_object)
}
\arguments{
\item{lm_analysis_object}{an object of class 'lm_analysis' created e.g. by \code{\link{make_lm}}.}
}
\value{
the data frame used for modeling.
}
\description{
The model frame of a lm_analysis object as specified by \code{\link[stats]{model.frame}}.
}
