% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop.R
\name{nobs}
\alias{nobs}
\title{Extract numbers of complete observations}
\usage{
nobs(x, ...)
}
\value{
The number of complete observations.
}
\description{
S3 generic functions used to extract the numbers of complete observations
}
