% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_funs.R
\name{normality}
\alias{normality}
\title{Normality of the model residuals}
\usage{
normality(linear_model, ...)
}
\arguments{
\item{linear_model}{lm, glm, polr or lm_analysis class model.}

\item{...}{extra arguments passed to \code{\link[broom]{augment}} and \code{\link{get_qc_tbl}}.}
}
\value{
a tibble with the test results: test statistic and p value.
}
\description{
Tests normality of the model residuals by Shapiro-Wilk test.
}
