% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{outer_rbind}
\alias{outer_rbind}
\title{Bind two data frames differing in column names by rows.}
\usage{
outer_rbind(tbl1, tbl2)
}
\arguments{
\item{tbl1}{data frame.}

\item{tbl2}{data frame.}
}
\value{
a data frame bound by rows.
}
\description{
binds two data frames with different sets of variables.
The non-common ones are filled with NAs.
}
