% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_forest.default}
\alias{plot_forest.default}
\title{Display modeling results in a Forest plot.}
\usage{
\method{plot_forest}{default}(
  data,
  variable = "variable",
  level = "level",
  confounder = "confounder",
  n = "n",
  n_complete = "n_complete",
  estimate = "estimate",
  lower_ci = "lower_ci",
  upper_ci = "upper_ci",
  p_value = "p_value",
  plot_title = NULL,
  plot_subtitle = NULL,
  plot_tag = NULL,
  x_lab = expression(beta * ", 95\% CI"),
  point_size = 2,
  point_shape = 16,
  label_estimate = TRUE,
  estimate_size = 2.75,
  estimate_hjust = 0.2,
  estimate_vjust = -0.7,
  hide_confounder = TRUE,
  hide_baseline = FALSE,
  baseline_label = "Baseline",
  signif_digits = 2,
  x_text_separator = "\\n",
  cutpoint = 0,
  x_axis_trans = "identity",
  cust_theme = ggplot2::theme_classic()
)
}
\arguments{
\item{data}{data frame.}

\item{variable}{name of the variable storing explanatory feature labels.}

\item{level}{name of the variable with storing the labels of levels for categorical explanatory features.}

\item{confounder}{name of a text variable (coded yes/no) indicating if the explanatory feature has to be treated as a confounder
and optionally skipped from the plot.}

\item{n}{name of the variable storing the n numbers for the levels of categorical explanatory features.}

\item{n_complete}{name of the variable storing the number of complete observations.}

\item{estimate}{name of the variable with the estimate values.}

\item{lower_ci}{name of the variable with the lower CI values.}

\item{upper_ci}{name of the variable with the upper CI values.}

\item{p_value}{name of the variable storing p values.}

\item{plot_title}{text displayed as a plot title.}

\item{plot_subtitle}{text displayed as a plot subtitle.}

\item{plot_tag}{text displayed as a plot tag.}

\item{x_lab}{text displayed as a X axis title.}

\item{point_size}{size of the points (coding for estimate values) in the Forest plot.}

\item{point_shape}{shape of the points (coding for estimate values) in the Forest plot.}

\item{label_estimate}{logical, should estimate value with 95\% CI be displayed in the plot?}

\item{estimate_size}{size of the estimate label.}

\item{estimate_hjust}{horizontal justification of the estimate label.}

\item{estimate_vjust}{vertical justification of the estimate label.}

\item{hide_confounder}{logical, should confounder features be displayed in the plot?}

\item{baseline_label}{text to be displayed in the Y axis for the baseline/model intercept.}

\item{signif_digits}{significant digits, used for rounding of the estimate and 95\% CI values.}

\item{x_text_separator}{separator of the Y axis text labels.}

\item{cutpoint}{regulation cutoff displayed in the plots as a dashed vertical line.}

\item{x_axis_trans}{transformation of the X axis, as specified for \code{\link[ggplot2]{scale_x_continuous}}.}

\item{cust_theme}{custom ggplot2 theme.}
}
\value{
a Forest plot as a ggplot object.
}
\description{
Generates a customized Forest plot with the model estimates and 95\% confidence intervals (CI).
}
\details{
Designed to work optimally with the output of the \code{\link{summary.lm_analysis}} method applied
to the objects of 'lm_analysis' class generated e.g. by \code{\link{make_lm}}.
}
