% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_forest.lm_analysis}
\alias{plot_forest.lm_analysis}
\title{Display modeling results of a lm_anaylsis object in a Forest plot.}
\usage{
\method{plot_forest}{lm_analysis}(
  lm_analysis_object,
  plot_title = NULL,
  stats_position = c("subtitle", "tag", "none"),
  show_stats = c("n_complete", "adj_rsq", "aic", "mae"),
  cust_theme = ggplot2::theme_classic(),
  signif_digits = 2,
  ...
)
}
\arguments{
\item{lm_analysis_object}{an object of class 'lm_analysis' created e.g. by \code{\link{make_lm}}.}

\item{plot_title}{text to be displayed in the plot title.}

\item{stats_position}{position, where the fit stats are displayed, subtitle by default.}

\item{show_stats}{fit stats to be presented in the plot, see \code{\link{summary.lm_analysis}} for details.}

\item{cust_theme}{cust_theme custom ggplot2 theme.}

\item{signif_digits}{significant digits, used for rounding of the estimate, 95\% CI values and fit stats.}

\item{...}{extra arguments passed to \code{\link{plot_forest.default}}.}
}
\description{
Generates a customized Forest plot with the model estimates and 95\% confidence intervals (CI).
}
\details{
A method specific method for lm_analysis object (constructed e.g. by \code{\link{make_lm}}).
Generates a ggplot with estimate and CI values as described for \code{\link{plot_forest.default}}
and presents extra model statistics in the sub-title or tag.
}
