% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_forest}
\alias{plot_forest}
\title{The generic function used for drawing of Forest plots.}
\usage{
plot_forest(x, ...)
}
\arguments{
\item{x}{object providing data for drawing such as estimates, confidence intervals and variable names.}

\item{...}{extra arguments passed to specific methods.}
}
\description{
Draws a Forest plot given a data frame or lm_analysis object.
}
