% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_importance}
\alias{plot_importance}
\title{Plot percentage of variance/deviance for lm_analysis objects.}
\usage{
plot_importance(
  lm_analysis_object,
  plot_form = c("bar", "scree"),
  plot_title = NULL,
  plot_subtitle = NULL,
  plot_tag = NULL,
  cust_theme = ggplot2::theme_classic(),
  bar_color = "steelblue",
  line_color = "gray70",
  line_size = 0.5,
  show_points = TRUE,
  point_color = "steelblue",
  point_size = 2,
  residual_color = "gray60",
  show_p = TRUE,
  p_size = 2.75,
  p_hjust = -0.2,
  p_vjust = 0,
  signif_digits = 2,
  ...
)
}
\arguments{
\item{lm_analysis_object}{an object of class 'lm_analysis' created e.g. by \code{\link{make_lm}}.}

\item{plot_form}{form of the plot, defaults to bar.}

\item{plot_title}{text presented in the plot title.}

\item{plot_subtitle}{text presented in the plot subtitle.}

\item{plot_tag}{text presented in the plot tag.}

\item{cust_theme}{custom ggplot2 theme.}

\item{bar_color}{bar fill color, ignored for the scree plot option.}

\item{line_color}{line color, ignored for the bar plot option.}

\item{line_size}{line size, ignored for the bar plot option.}

\item{show_points}{logical, should point be displayed in the scree plot? ignored for the bar plot option.}

\item{point_color}{point color, ignored it show_points is FALSE, ignored for the bar plot option.}

\item{point_size}{point size, ignored it show_points is FALSE, ignored for the bar plot option.}

\item{residual_color}{color of the residual bar/point.}

\item{show_p}{logical, should ANOVA p values be displayed in the plot?}

\item{p_size}{size of the p value text, ignored if show_p is FALSE.}

\item{p_hjust}{horizontal justification of the p value text, ignored if show_p is FALSE.}

\item{p_vjust}{vertical justification of the p value text, ignored if show_p is FALSE.}

\item{signif_digits}{significant digits, used for rounding of the p value.}

\item{...}{additional arguments passed to ANOVA calculating functions, see: \code{\link{anova.lm_analysis}}.}
}
\value{
a bar or scree ggplot graphic.
}
\description{
Presents percent data set variance/deviance explained by the model components
as a bar or scree plot.
}
\details{
Percentage of variance/deviance is calulated as described for \code{\link{anova.lm_analysis}}.
}
