% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{point_plot_}
\alias{point_plot_}
\title{Generate a customized point plot.}
\usage{
point_plot_(
  data,
  x_var,
  y_var,
  x_lab = x_var,
  y_lab = y_var,
  plot_title = NULL,
  plot_subtitle = NULL,
  plot_tag = NULL,
  smooth = TRUE,
  silent = TRUE,
  ...
)
}
\arguments{
\item{data}{data frame.}

\item{x_var}{name of the variable to be shown in the x axis.}

\item{y_var}{name of the variable to be shown in the y axis.}

\item{x_lab}{x axis title.}

\item{y_lab}{y axis title.}

\item{plot_title}{plot title.}

\item{plot_subtitle}{plot subtitle.}

\item{plot_tag}{plot tag.}

\item{smooth}{logical, should a trend line be displayed.}

\item{silent}{logical, display warnings?}

\item{...}{extra arguments passed to geom_smooth().}
}
\value{
a ggplot graphic
}
\description{
draws a simple point plot for model diagnostic purposes.
}
\details{
draws a simple point plot for diagnostic purposes. As per design, takes the output
of get_qc_tbl() as data argument, color-codes model potential outliers.
}
