% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop.R
\name{predict.lm_analysis}
\alias{predict.lm_analysis}
\title{Predictions for a lm_analysis model.}
\usage{
\method{predict}{lm_analysis}(object, ...)
}
\arguments{
\item{object}{an object of class 'lm_analysis' created e.g. by \code{\link{make_lm}}.}

\item{...}{extra arguments passed to specific prediction methods such as \code{\link[stats]{predict.lm}}
or \code{\link[stats]{predict.glm}}.}
}
\value{
a vector or data frame with the predicted values
}
\description{
Predictions for a lm_analysis model.
}
