% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop.R
\name{print.lm_analysis}
\alias{print.lm_analysis}
\title{Prints the base properties of a lm_analysis object.}
\usage{
\method{print}{lm_analysis}(x)
}
\arguments{
\item{x}{an object of class 'lm_analysis' created e.g. by \code{\link{make_lm}}.}
}
\value{
Nothing, called for side effects.
}
\description{
Prints a handy summary of a lm_analysis object.
}
