% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_funs.R
\name{prop_odds}
\alias{prop_odds}
\title{Test odds proportionality assumption for polr models}
\usage{
prop_odds(linear_model, ...)
}
\arguments{
\item{linear_model}{polr model of lm_analysis class model.}

\item{...}{extra arguments passed to \code{\link[brant]{brant}}.}
}
\value{
a tibble with the test results: test statistic, degrees of freedom and p value.
}
\description{
Tests odds proportionality assumption for ordinal regression models.
An implementation of \code{\link[brant]{brant}}.
}
