% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop.R
\name{residuals.lm_analysis}
\alias{residuals.lm_analysis}
\title{Residuals of a lm_analysis model.}
\usage{
\method{residuals}{lm_analysis}(lm_analysis_object, ...)
}
\arguments{
\item{lm_analysis_object}{an object of class 'lm_analysis' created e.g. by \code{\link{make_lm}}.}

\item{...}{extra arguments passed to \code{\link{get_qc_tbl}}.}
}
\value{
a data frame with predicted values and residuals.
}
\description{
Extracts enhanced residuals for a lm_analysis model.
}
