% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop.R
\name{step.default}
\alias{step.default}
\title{Default method for AIC driven elimination.}
\usage{
\method{step}{default}(object, ...)
}
\arguments{
\item{object}{model object.}

\item{...}{other arguments passed to \code{\link[stats]{step}}.}
}
\description{
Applies \code{\link[stats]{step}} to the lm_analysis object.
}
