% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop.R
\name{step.lm_analysis}
\alias{step.lm_analysis}
\title{Model optimization by AIC driven elimination}
\usage{
\method{step}{lm_analysis}(lm_analysis_object, ...)
}
\arguments{
\item{lm_analysis_object}{an object of class 'lm_analysis' created e.g. by \code{\link{make_lm}}.}

\item{...}{other arguments passed to \code{\link[stats]{step}}.}
}
\value{
A lm_analysis object with the reduced set of explanatory variables.
}
\description{
Applies \code{\link[stats]{step}} to the lm_analysis object.
}
