% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop.R
\name{step}
\alias{step}
\title{Optimize models by AIC-driven elimination.}
\usage{
step(x, ...)
}
\arguments{
\item{x}{model object.}
}
\value{
A model of the same class as x with the optimized set of explanatory variables.
}
\description{
S3 generic functions used to optimize models by AIC-driven term elimination
}
