% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop.R
\name{summary.lm_analysis}
\alias{summary.lm_analysis}
\title{Summarize inference or fit stats for a lm_analysis object.}
\usage{
\method{summary}{lm_analysis}(lm_analysis_object, type = c("inference", "fit"), ...)
}
\arguments{
\item{lm_analysis_object}{an object of class 'lm_analysis' created e.g. by \code{\link{make_lm}}.}

\item{type}{type of the statistics returned.}

\item{...}{extra arguments passed to the \code{\link{get_estimates}} or \code{\link{get_stats}} functions.}
}
\value{
a data frame with requested statistisc as specified for \code{\link{get_estimates}} or \code{\link{get_stats}}.
}
\description{
Retrieves inference or fit statistics from an lm_analysis object.
}
