% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_funs.R
\name{get_estimates}
\alias{get_estimates}
\title{Extract model estimates with inference statistics.}
\usage{
get_estimates(
  linear_model,
  transf_fun = identity,
  ci_method = c("default", "distribution", "normal"),
  silent_messages = TRUE,
  ...
)
}
\arguments{
\item{linear_model}{lm or glm class model or an object of the lm_analysis class.}

\item{transf_fun}{function used for transformation of the estimates and confidence intervals, identity() by default}

\item{ci_method}{specifies how the confidence intervals should be calculated, see details.}

\item{silent_messages}{logical, silence the messages returned during confidence interval calculation}

\item{...}{extra arguments passed to \code{\link[stats]{confint}}}
}
\value{
a data frame with the parameter names (as specified by the summary() generic function),
numbers of complete observations used for model construction (n_complete),
transformed beta estimate and confidence interval values (lower_ci, upper_ci)
and p value
}
\description{
extracts coefficients from a linear or generalized linear model (beta) with 95\% confidence intervals (CI)
and p values.
}
\details{
the output estimates and ci are transformed with the 'transf_fun' argument.
The ci_method option: calculates the CI per \code{\link[stats]{confint}} method by default, from the t or normal distribution
as appropriate for the model type ('distribution') or explicitly from the normal distribution ('normal').
based on the normality assumption: i. e. SEM * critical normal distribution value.
No correction for multiple testing is applied to the beta significance testing.
}
