% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_funs.R
\name{get_stats}
\alias{get_stats}
\title{Extract model quality statistics.}
\usage{
get_stats(
  linear_model,
  rsq_type = c("v", "kl", "sse", "lr", "n"),
  type.residuals = c("working", "response", "deviance", "pearson", "partial")
)
}
\arguments{
\item{linear_model}{lm or glm class model or an object of the lm_analysis class.}

\item{type.residuals}{type of the residuals used for error calculation, defaults to 'working'.}

\item{r_sq_type}{the type of R-squared (only applicable for generalized linear models), please refer to \code{\link[rsq]{rsq}} for details.}
}
\value{
a data frame with the n number of complete observations, AIC, BIC, R squared, deviance, mae (mean absolute error),
mse (mean squared error)
}
\description{
calculates AIC, BIC, R squared, mean squared and absolute error of the given linear model
}
\details{
R squared is calculated as described for \code{\link[rsq]{rsq}}
}
