% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop.R
\name{nobs.lm_analysis}
\alias{nobs.lm_analysis}
\title{Number of complete observations in a lm_analysis object}
\usage{
\method{nobs}{lm_analysis}(x, ...)
}
\arguments{
\item{x}{an object of class 'lm_analysis' created e.g. by \code{\link{make_lm}}.}
}
\value{
number of complete cases used for modeling.
}
\description{
The number of complete observations as returned by the \code{\link[stats]{nobs}} generic.
}
