% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop.R
\name{plot.lm_analysis}
\alias{plot.lm_analysis}
\title{Generates diagnostic plots for a lm_analysis object.}
\usage{
\method{plot}{lm_analysis}(lm_analysis_object, cust_theme = ggplot2::theme_classic(), ...)
}
\arguments{
\item{lm_analysis_object}{an object of class 'lm_analysis' created e.g. by \code{\link{make_lm}}.}

\item{cust_theme}{customized plot theme provided by the user.}

\item{...}{extra arguments passed to \code{\link[broom]{augment}} and \code{\link{get_qc_tbl}}.}
}
\value{
returns a series of ggplot objects with the diagnostic residuals plots.
}
\description{
Makes a series of plots of the model residuals for visual quality control.
}
